requirements.txt
Django==5.2.5
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
weasyprint==66.0
Pillow==11.3.0
asgiref==3.9.1
sqlparse==0.5.3

